camera_scheduler/pyproject.toml
[project]
name = "dahua-asi"
version = "0.1.0"
description = "Python client and CLI for Dahua DHI-ASI7213S-W access control terminals (RPC2 protocol)."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "bekzod" }]
license = { text = "MIT" }
dependencies = [
    "requests>=2.31",
    "click>=8.1",
    "python-dotenv>=1.0",
    "amcrest>=1.9.9",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.49",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "types-requests",
]

[project.scripts]
dahua-logs = "dahua_asi.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dahua_asi"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
